+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
   - Fixes a bug that caused the Omeka item search feature in the record form to display the the title of an item's parent _collection_, not the item itself, when the item belonged to a collection.
 
+  - In the editor, when a modal window is open (eg, a record delete confirmation pop-up) and the route is changed, the modal is now automatically closed. 
+
 ## v2.1.3 ([commits](https://github.com/scholarslab/Neatline/compare/2.1.2...2.1.3)) ~ October 1, 2013
 
 #### Added Features